requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.13.1
mypy==0.931
mypy-extensions==0.4.3
Pillow==9.0.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
types-pytz==2021.3.4
types-PyYAML==6.0.3
typing_extensions==4.0.1
tzdata==2021.5
